LayoutTests/media/track/track-css-matching-lang-expected.txt

Test that cues are being matched properly by the lang attribute and :lang() pseudo class.
EVENT(canplaythrough)
EVENT(seeked)
EXPECTED (getComputedStyle(cueNode).color == 'rgb(128, 0, 128)') OK
EXPECTED (getComputedStyle(cueNode).color == 'rgb(0, 255, 0)') OK
EXPECTED (getComputedStyle(cueNode).color == 'rgb(128, 0, 128)') OK
END OF TEST
